cyclical error correction, calibration
