+++ Cadik08/common/math.h
@@ -6,9 +6,9 @@ namespace com
 namespace math
 {
 
-unsigned ceil_multiple(const unsigned, const unsigned short);
-unsigned floor_multiple(const unsigned, const unsigned short);
-unsigned ceil_power(const unsigned, const unsigned short);
+unsigned ceil2mul(const unsigned, const unsigned short);
+unsigned floor2mul(const unsigned, const unsigned short);
+unsigned ceil2pow(const unsigned, const unsigned short);
 
 }// namespace math
 }// namespace com
