+++ Cadik08/doc/report.bib
@@ -1,0 +1,20 @@
+@InProceedings{cadik07color_to_gray,
+  author =     {Laszlo Neumann and Martin {\v{C}}ad\'{i}k and Antal Nemcsics},
+  title =      {An Efficient Perception-Based Adaptive Color to Gray Transformation},
+  booktitle =  {Proceedings of Computational Aesthetics 2007},
+  pages =      {73-- 80},
+  publisher =  {Eurographics Association},
+  year =       {2007},
+  address =    {Banff, Canada},
+}
+
+@inproceedings{gooch2005color2gray,
+  title={Color2gray: salience-preserving color removal},
+  author={Gooch, Amy A and Olsen, Sven C and Tumblin, Jack and Gooch, Bruce},
+  booktitle={ACM Transactions on Graphics (TOG)},
+  volume={24},
+  number={3},
+  pages={634--639},
+  year={2005},
+  organization={ACM}
+}
